requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg[binary]
python-dotenv
pydantic
email-validator
python-jose[cryptography]
passlib[bcrypt]
python-multipart
